improvements to the timeline protocol parsing
